requirements/base.txt
boto3==1.4.4
Django==1.11.1
django-boto==0.3.11
django-bootstrap3==8.2.1
django-countries==4.2
django-debug-toolbar==1.7
django-extensions==1.7.7
django-haystack==2.6.0
django-phonenumber-field==1.2.0
django-smart-selects==1.5.2
django-storages-redux==1.3.2
Pillow==4.0.0
psycopg2==2.7
